Travel time assignment

+++ week_2/exercise_2_14_8.py
@@ -1,5 +1,5 @@
 #2.8
-time_now = int(input("What time is it now? (In 24 hour format)"))
-timer_duration = int(input("How long do you want a timer for?"))
-time_end = ((timer_duration + time_now) % 24)
+time_now = float(input("What time is it now? (In 24 hour format)"))
+timer_duration = float(input("How long do you want a timer for?"))
+time_end = ((timer_duration + time_now) % 24.00)
 print("It will be", time_end, "O' Clock")
